move server input to shared
